Fix additional build errors
- PerformanceOptimizations.cs: Cast double to decimal for cost calculation
- ErrorHandling.cs: Add null check for path traversal validation
- HighStakesFeatures.cs: Fix TapeVerificationResult to use Errors list
- HighStakesFeatures.cs: Fix ShareData references to use Data property
- ObservabilityEnhancements.cs: Fix Channel type to DistributedTraceSpan
- ObservabilityEnhancements.cs: Fix StartSpan to return DistributedTraceSpan
- OpenTelemetryIntegration.cs: Fix DateTime.UtcDateTime to just DateTime
- OpenTelemetryIntegration.cs: Fix metric value access (CounterMetric/GaugeMetric)

+++ DataWarehouse.SDK/Infrastructure/PerformanceOptimizations.cs
@@ -1274,7 +1274,7 @@ public sealed class TieredStorageManager : IAsyncDisposable
             TierStatistics = tierStats,
             TotalObjects = _locationIndex.Count,
             TotalBytes = _locationIndex.Values.Sum(l => l.Size),
-            EstimatedMonthlyCost = tierStats.Sum(t => t.TotalBytes / (1024.0 * 1024 * 1024) * t.CostPerGbMonth)
+            EstimatedMonthlyCost = tierStats.Sum(t => (decimal)(t.TotalBytes / (1024.0 * 1024 * 1024)) * t.CostPerGbMonth)
         };
     }
